Clear CFLAGS before configure.
Previously default autoconf values were getting used.
This change drops those defaults.

All flags decided by us at configure are now in AM_CFLAGS and
that is substituted into the Makefile.

If a user wants to add to the cflags use EXTRA_CFLAGS when building (not
configuring).

If a user wants to change cflags use CFLAGS when configuring.